### 2019-09-19 #1747 @drvinceknight
Hi @inakleinbottle, [nice to see you at PyCon UK!] how do you feel about possibly reviewing this paper for JOSS?

Hi @virgesmith @jaroslavhron would you be interested in reviewing this paper?

It is presenting an implementation of two algorithms for solutions of [a particular type of] sparse linear systems.

If anyone has any questions about reviewing for JOSS there is documentation here
