### 2019-09-23 #1747 @inakleinbottle
I have completed my review, and opened an [issue](https://github.com/GeoStat-Framework/pentapy/issues/1#issue-497003090) on the submission repository detailing some issues that need to be addressed.

(Do I need to open a Review issue using Whedon?)
