### 2019-09-23 #1747 @virgesmith
hi, sorry I just assumed one reviewer (that has been my experience of JOSS) and @inakleinbottle  beat me to it. Happy to review, but very unlikely I could submit my report before the end of the month, if this is OK
