### 2019-09-24 #1759 @whedon
[REVIEW]: pentapy: A Python toolbox for pentadiagonal linear systems
**Submitting author:** @MuellerSeb (<a href="http://orcid.org/0000-0001-9060-4008">Sebastian Müller</a>)
**Repository:** <a href="https://github.com/GeoStat-Framework/pentapy" target ="_blank">https://github.com/GeoStat-Framework/pentapy</a>
**Version:** v1.0.2
