### 2019-09-24 fac2fbf
blackened docs conf [skip ci]
