+++ docs/source/conf.py
@@ -208,9 +208,7 @@ latex_documents = [
 
 # One entry per manual page. List of tuples
 # (source start file, name, description, authors, manual section).
-man_pages = [
-    (master_doc, "pentapy", "pentapy Documentation", [author], 1)
-]
+man_pages = [(master_doc, "pentapy", "pentapy Documentation", [author], 1)]
 
 
 # -- Options for Texinfo output -------------------------------------------
