### 2019-09-24 f5faf19
Use warnings for Cython and ZeroDivision when using PTRANS algorithms
