### 2019-09-24 8a8beb7
Setup.py: add extra arguments for scipy and umf solvers
