### 2019-09-24 cc2ffcc
Examples: Add example for PTRANS-I corner case; Add doc-strings
