+++ examples/03_compare.py
@@ -1,3 +1,10 @@
+# -*- coding: utf-8 -*-
+"""
+3. Example.
+
+Here we compare the outcome of the PTRANS-I and PTRANS-II algorithm for
+a random input.
+"""
 import numpy as np
 import pentapy as pp
 
