### 2019-09-24 c9e6bf8
Add CONTRIBUTING.md [skip ci]
