### 2019-09-24 7f3695d
update README and rtfd [skip ci]
