### 2019-09-24 e75e99d
update test, check corner case, check warning
