### 2019-09-24 2a8f954
allways show warnings for CI in unittest
