### 2019-09-24 286948b
check warning test only if there is a warning for CI in unittest
