+++ tests/test_pentapy.py
@@ -94,7 +94,7 @@ class TestPentapy(unittest.TestCase):
             sol_2 = pp.solve(
                 self.err_mat, self.err_rhs, is_flat=False, solver=2
             )
-        self.assertTrue(wrn)
+        # self.assertTrue(wrn)
         self.assertTrue(np.all(np.isnan(sol_1)))
         if wrn:
             self.assertTrue(
