### 2019-09-24 ddc470c
check nan only if there is a warning for CI in unittest
