### 2019-09-24 #1759 @MuellerSeb
My response to @inakleinbottle can be found here: https://github.com/GeoStat-Framework/pentapy/issues/1#issuecomment-534624645
