### 2019-09-29 GeoStat-Framework/pentapy#2 @virgesmith
JOSS review
# Overview

This is a neat and well-defined package that provides a big performance improvement on existing `python` algorithms for solving linear systems of pentadiagonal matrices, and I found it to work as claimed. My main reservation is the documentation, which I think could be improved somewhat, but if the points I make below are addressed/defended satisfactorily, I would be happy to accept it for publication in JOSS.

# Installation

The package installed correctly via pip. Dependencies are dealt with by setup.py but `scipy` (which is required for some of the algorithms) is not installed automatically, and has to be installed manually. Perhaps this could be automatic, and/or see comment in Documentation section below.

# Software

For solver=1,2 there is no check for singular matrix so you get a `ZeroDivisionError: float division`. Algorithms 3,4,5 report this situation specifically (as either an error or a warning). If performance is not impacted, could you check for this? Or simply catch numerical errors and re-raise with a more suggestive error message(s).

# Tests

I found the solver to work correctly on some basic but inexhaustive tests of my own - discretising the heat equation both tri- and pentadiagonally and comparing the result to an analytic solution.

I think there should be more tests than are currently implemented in the test suite, and they should include some more realistic examples as well as the random matrix tests. There should also be some basic unit tests of the matrix utilities, e.g. check full->banded->full gives the original matrix.

Documentation on how to run (and extend) the package tests should be provided in the community guidelines, if the author is willing to accept contributions.

# Performance

My tests agree with the claims in the paper. But the code only supports 5 of the 7 algorithms compared in the graphs in the README/readthedocs, could those graphs be made consistent with the one in the paper?

# Documentation

The README looks a little rushed, e.g. "pentapy is a toolbox to deal with pentadiagonal matrices in Python" is a bit vague and to a large extent the README just duplicates the readthedocs documentation. Personally, I would have a README containing only a statement of intent, authorship and community guidelines, basic installation instructions, and links to detailed documentation and examples (in readthedocs).

Re: community guidelines,  I don't see "clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support".

Whilst `help(pentapy)` provides API-level documentation (some typos/spelling issues here), I think this should also be in the main documentation, the `solve` and `create_full` methods are only dealt with in the one example given. Other methods: `diag_indices`, `shift_banded`, `create_banded` are not mentioned at all.

The Installation and Requirements sections of the README could be merged and made clearer, esp. that `scipy` is an optional dependency that require.

## Paper

I think the "unique selling point" of this package is a very siginificant improvement in performance over current implementations/algorithms, and this should be emphasised more in the paper (and the README).
