### 2019-09-30 GeoStat-Framework/pentapy#1 @inakleinbottle
Thank you for addressing these points. I've updated the review issue. I have two points outstanding, one of which just needs some clarification from the editor (citations for packages mentioned in the paper).
