### 2019-09-30 #1759 @inakleinbottle
I have a query for the editor: copied from my review
* You might need to include references (at the end of the paper) to the packages mentioned in your paper: numpy, scipy, Cython, perfplot. This needs to be clarified by the JOSS editors.

What is the policy regarding citation of software packages in references? I should comment that these are clearly acknowledged in the paper, but are not included in the list of references.
