### 2019-09-30 #1759 @drvinceknight
Hi @inakleinbottle, good question, I note that `scipy` has been referenced however I feel that on this occasion references for `cython` and `perfplot` would be helpful for the reader (for example reading the paper myself I was not familiar with `perfplot`).

I do not feel strongly about including a reference for `numpy` in the context of the text of the paper but I would suggest that it could be nice to include it for completeness.

Does that sound ok to you @MuellerSeb ?
