### 2019-09-30 #1759 @MuellerSeb
@drvinceknight : Thank you for working out a way to go. I will include a reference to the github repositories of numpy and perfplot (numpy has no "real" code-publication and perfplot is a rather small package without any publications). In case of Cython, a dedicated publication exists.
