### 2019-09-30 27e62bc
Update tests for warnings (only on 64bit)
