+++ tests/test_pentapy.py
@@ -4,6 +4,7 @@ This is the unittest for pentapy.
 """
 from __future__ import division, absolute_import, print_function
 
+import platform
 import warnings
 import unittest
 import numpy as np
@@ -86,21 +87,20 @@ class TestPentapy(unittest.TestCase):
             [[3, 2, 1, 0], [-3, -2, 7, 1], [3, 2, -1, 5], [0, 1, 2, 3]]
         )
         self.err_rhs = np.array([6, 3, 9, 6])
-        # https://stackoverflow.com/a/32089134/6696397
-        with warnings.catch_warnings(record=True) as wrn:
-            sol_1 = pp.solve(
-                self.err_mat, self.err_rhs, is_flat=False, solver=1
-            )
-            sol_2 = pp.solve(
-                self.err_mat, self.err_rhs, is_flat=False, solver=2
-            )
-        # self.assertTrue(wrn)
-        if wrn:
+        # there is a cython bug for 32bit where no ZeroDivisionError is raised
+        if platform.architecture()[0] == "64bit":
+            # https://stackoverflow.com/a/32089134/6696397
+            with warnings.catch_warnings(record=True) as wrn:
+                sol_1 = pp.solve(
+                    self.err_mat, self.err_rhs, is_flat=False, solver=1
+                )
+            self.assertTrue(wrn)
             self.assertTrue(np.all(np.isnan(sol_1)))
             self.assertTrue(
                 str(wrn[0].message)
                 == "pentapy: PTRANS-I not suitable for input-matrix."
             )
+        sol_2 = pp.solve(self.err_mat, self.err_rhs, is_flat=False, solver=2)
         diff_2 = np.max(np.abs(np.dot(self.err_mat, sol_2) - self.err_rhs))
         self.assertAlmostEqual(diff_2, 0.0)
 
