+++ examples/05_perform_full_mat.py
@@ -3,7 +3,7 @@
 5. Example.
 
 Here we compare all algorithms for solving pentadiagonal systems provided
-by pentapy (except umf) and the standard numpy linear algebra solver.
+by pentapy (except umf) using a full quadratic matrix as input.
 
 To use this script you need to have the following packages installed:
 
