### 2019-09-30 eed9436
add more tests for tools
