### 2019-09-30 74de4c2
simpler performance plot in README
