+++ README.md
@@ -33,11 +33,11 @@ pentapy installed, you have to provide a c-compiler and run:
     pip install pentapy
 
 To get the scipy solvers running, you have to install scipy or you can use the
-extra argument:
+following extra argument:
 
     pip install pentapy[all]
 
-Instead of "all" you can also typ "scipy" or "umfpack".
+Instead of "all" you can also typ "scipy" or "umfpack" to get one of these specific packages.
 
 
 ## References
@@ -78,22 +78,23 @@ This should give something like:
 
 ### Performance
 
-In the following a couple of solvers for pentadiagonal systems is compared:
+In the following a couple of solvers for pentadiagonal systems are compared:
 
-* Solver 1: Python implementation of ``pentapy.solve``
-* Solver 2: Cython implementation of ``pentapy.solve``
-* Solver 3: [``scipy.sparse.linalg.spsolve``](http://scipy.github.io/devdocs/generated/scipy.sparse.linalg.spsolve.html) with ``use_umfpack=False``
-* Solver 4: [``scipy.sparse.linalg.spsolve``](http://scipy.github.io/devdocs/generated/scipy.sparse.linalg.spsolve.html) with [``use_umfpack=True``](https://scikit-umfpack.github.io/scikit-umfpack/)
-* Solver 5: [Lapack solver](http://www.netlib.org/lapack/explore-html/d3/d49/group__double_g_bsolve_gafa35ce1d7865b80563bbed6317050ad7.html) for diagonal matrices [``scipy.linalg.lapack.dgbsv``](scipy.github.io/devdocs/generated/scipy.linalg.lapack.dgbsv.html)
-* Solver 6: Scipy banded solver [``scipy.linalg.solve_banded``](scipy.github.io/devdocs/generated/scipy.linalg.solve_banded.html)
-* Solver 7: Standard solver of Numpy [``np.linalg.solve``](https://www.numpy.org/devdocs/reference/generated/numpy.linalg.solve.html)
+* Solver 1: Standard linear algebra solver of Numpy [``np.linalg.solve``](https://www.numpy.org/devdocs/reference/generated/numpy.linalg.solve.html)
+* Solver 2: [``scipy.sparse.linalg.spsolve``](http://scipy.github.io/devdocs/generated/scipy.sparse.linalg.spsolve.html)
+* Solver 3: Scipy banded solver [``scipy.linalg.solve_banded``](scipy.github.io/devdocs/generated/scipy.linalg.solve_banded.html)
+* Solver 4: pentapy.solve with ``solver=1``
+* Solver 5: pentapy.solve with ``solver=2``
 
 <p align="center">
-<img src="https://raw.githubusercontent.com/GeoStat-Framework/pentapy/master/examples/perfplot.png" alt="Performance" width="600px"/>
+<img src="https://raw.githubusercontent.com/GeoStat-Framework/pentapy/master/examples/perfplot_simple.png" alt="Performance" width="600px"/>
 </p>
 
+The implementations of pentapy are almost one order of magnitude faster than the
+scipy algorithms for banded or sparse matrices.
+
 The performance plot was created with [``perfplot``](https://github.com/nschloe/perfplot).
-Have a look at the script: [``examples/02_perform.py``](https://github.com/GeoStat-Framework/pentapy/blob/master/examples/02_perform.py).
+Have a look at the script: [``examples/04_perform_simple.py``](https://github.com/GeoStat-Framework/pentapy/blob/master/examples/04_perform_simple.py.py).
 
 
 
