### 2019-09-30 bae6ae1
tests: use numpy to check array equality
