### 2019-09-30 f71648b
tests: better numpy array comparison
