### 2019-09-30 f1b18de
DOC-string update
