### 2019-09-30 868bead
update paper (perfplot, references) [skip ci]
