+++ paper/paper.bib
@@ -1,27 +1,20 @@
-@Misc{scipy,
-  author =    {Eric Jones and Travis Oliphant and Pearu Peterson and others},
-  title =     {{SciPy}: Open source scientific tools for {Python}},
-  year =      {2001--2019},
-  url = "http://www.scipy.org/",
-  note = {[Online; accessed 2019-06-28]}
-}
-
 @misc{zenodo,
   author       = {Sebastian Müller},
   title        = {pentapy},
-  year         = 2019,
+  year         = {2019},
+  journal      = {Zenodo},
   doi          = {10.5281/zenodo.2587158},
   url          = {https://doi.org/10.5281/zenodo.2587158}
 }
 
 @article{askar,
-  title={On solving pentadiagonal linear systems via transformations},
-  author={Askar, SS and Karawia, AA},
-  journal={Mathematical Problems in Engineering},
-  volume={2015},
-  year={2015},
-  doi={10.1155/2015/232456},
-  publisher={Hindawi}
+    title = {On solving pentadiagonal linear systems via transformations},
+    author = {Askar, SS and Karawia, AA},
+    journal = {Mathematical Problems in Engineering},
+    volume = {2015},
+    year = {2015},
+    doi = {10.1155/2015/232456},
+    publisher = {Hindawi}
 }
 
 @book{aquifer_test,
@@ -48,4 +41,44 @@
     url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/hyp.10064},
     eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/hyp.10064},
     year = {2014}
+}
+
+@misc{scipy,
+  author =    {Eric Jones and Travis Oliphant and Pearu Peterson and others},
+  title =     {{SciPy}: Open source scientific tools for {Python}},
+  year =      {2001--2019},
+  url = "http://www.scipy.org/",
+  note = {[Online; accessed 2019-06-28]}
+}
+
+@misc{numpy,
+  author = {Travis Oliphant and others},
+  title = {NumPy},
+  year = {2019},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  howpublished = {\url{https://github.com/numpy/numpy}}
+}
+
+@misc{perfplot,
+  author = {Nico Schlömer},
+  title = {perfplot},
+  year = {2019},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  howpublished = {\url{https://github.com/nschloe/perfplot}}
+}
+
+@article{cython,
+    author={S. {Behnel} and R. {Bradshaw} and C. {Citro} and L. {Dalcin} and D. S. {Seljebotn} and K. {Smith}},
+    journal={Computing in Science Engineering},
+    title={Cython: The Best of Both Worlds},
+    year={2011},
+    volume={13},
+    number={2},
+    pages={31-39},
+    keywords={C language;numerical analysis;Python language extension;Fortran code;numerical loops;Cython language;programming language;Sparse matrices;Runtime;Syntactics;Computer programs;Programming;Python;Cython;numerics;scientific computing},
+    doi={10.1109/MCSE.2010.118},
+    ISSN={},
+    month={March}
 }
