### 2019-09-30 73578da
update optional dependencies; rtd frontpage update [skip ci]
