### 2019-10-01 7db6dd3
Add explenation for tests for contributors [skip ci]
