### 2019-10-01 #1759 @MuellerSeb
My response to @virgesmith can be found here: https://github.com/GeoStat-Framework/pentapy/issues/2#issuecomment-537114042
