### 2019-10-02 GeoStat-Framework/pentapy#2 @virgesmith
@MuellerSeb thanks for addressing all those points. I'm happy now. (My fault re: the API docs I missed them completely.)
