### 2019-10-02 #1759 @virgesmith
@drvinceknight I am happy to recommend this package now, @MuellerSeb has addressed all the points I raised.
