### 2019-10-02 #1759 @MuellerSeb
@virgesmith and @inakleinbottle thank you for your detailed reviews. That really helped to further improve the package! @drvinceknight thank you for being such a nice editor!
Thanks everybody!
