### 2019-10-02 #1759 @drvinceknight
@openjournals/joss-eics this paper is ready to be accepted (unless I've missed something - this is my first paper as editor so please let me know if I'm forgetting something).
