### 2019-10-02 #1759 @MuellerSeb
I will make a release of pentapy 1.0.1 that includes all changes to get a proper DOI.
This is in line with the paper, since Version 1.0 is mentioned there, so I think a newer PATCH release is ok.
(Maybe it could be a good hint for further papers to use a pre-released release candidate for the review)
