### 2019-10-02 4eaeae8
zenodo version bump 1.0.2
