### 2019-10-02 #1759 @labarba
👋 @drvinceknight — The pre-publication steps to request authors are:

- [ ] Please make a tagged release of your software, and list the version tag of the archived version here.
- [ ] Please list the Zenodo DOI of the archived version here.
