### 2019-10-02 #1759 @MuellerSeb
Zenodo fails to release the 1.0.1 version... uargh!
I created an Issue: https://github.com/zenodo/zenodo/issues/1878
BUT: the DOI in the paper references the "general" Zenodo publication, that means, it always directs to the latest version. Is that ok?
