### 2019-10-02 #1759 @labarba
The JOSS paper should link to the DOI for the _specific version_ or the Zenodo archive that corresponds to the software _as reviewed and improved_ for the JOSS publication (not the general DOI for all versions).

Also, note that the Zenodo archive DOI is displayed in the margin decorators of the first page, and we thus request that you do not add it in your list of references. (It will also be included in the metadata of the Crossref deposit for the paper.)
