### 2019-10-02 #1759 @MuellerSeb
OK, I will remove it from the list of references. Now we have to wait until someone from zenodo fixes my failed release. v1.0.2 was just to trigger another zenodo release (which again failed), but I will leave it that way.
