+++ pentapy/_version.py
@@ -1,2 +1,2 @@
 """Provide a central version."""
-__version__ = "1.0.2"
+__version__ = "1.0.3.dev0"
