### 2019-10-07 cf9b873
remove Zenodo reference from the paper bib
