+++ paper/paper.bib
@@ -1,12 +1,3 @@
-@misc{zenodo,
-  author       = {Sebastian Müller},
-  title        = {pentapy},
-  year         = {2019},
-  journal      = {Zenodo},
-  doi          = {10.5281/zenodo.2587158},
-  url          = {https://doi.org/10.5281/zenodo.2587158}
-}
-
 @article{askar,
     title = {On solving pentadiagonal linear systems via transformations},
     author = {Askar, SS and Karawia, AA},
