### 2019-10-07 #1759 @MuellerSeb
The DOI from Zenodo was now successfully created:
10.5281/zenodo.3474843
