### 2019-10-08 #1759 @labarba
I made some minor copy edits in https://github.com/GeoStat-Framework/pentapy/pull/3
Note that the paper has links to the source repository and Zenodo archive on the margin decorators of the first page, and we don't need that mentioned in the body text.

I wasn't sure what the abbreviated word was here:
> backward resp. forward substitution

And finally, you present a figure with a performance test: it would be nice if you said what hardware you used for the test!
