### 2019-10-08 64469c0
Paper: resp. issue solve; add sys specs
