### 2019-10-08 #1759 @MuellerSeb
@labarba I merged your edits, removed the abbreviation and added the sys specs.
