### 2019-10-08 9c4ed6e
copy edit
https://github.com/openjournals/joss-reviews/issues/1759
