### 2019-10-08 ed39b0a
Update paper.md
typo.
