### 2019-10-08 #1759 @labarba
I'm not seeing the PDF yet, but the DOI resolves. I'll leave this issue open for now...
