### 2019-10-08 #1759 @labarba
👋 @arfon —The PDF is not showing in the embedded frame for me, even though the file does exist ... see: https://joss.theoj.org/papers/10.21105/joss.01759
